telegram_bot_PY/requirements.txt
aiogram
sqlite3
pytest
black
flake8
python-dotenv
